streamline cabin data handling and implement delete functionality

+++ src/services/apiCabins.js
@@ -1,48 +1,51 @@
-import supabase, { supabaseUrl } from "./Supabase";
+import supabase, { supabaseUrl } from "./supabase";
 
 export async function getCabins() {
   const { data, error } = await supabase.from("cabins").select("*");
 
   if (error) {
-    console.log(error);
-    throw new Error("Cabines couldn't be loaded");
+    console.error(error);
+    throw new Error("Cabins could not be loaded");
   }
+
   return data;
 }
 
-export async function deleteCabin(id) {
-  const { data, error } = await supabase.from("cabins").delete().eq("id", id);
-  if (error) {
-    console.error(error);
-    throw new Error("Cabin could not be deleted");
-  }
-  return data;
-}
-export async function createCabin(newCabin) {
+export async function createEditCabin(newCabin, id) {
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
+
   const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
     "/",
     ""
   );
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+  // 1. Create/edit cabin
+  let query = supabase.from("cabins");
 
-  // https://ckwcezaumxeiaghdilwu.supabase.co/storage/v1/object/public/cabin-images//cabin-001.jpg
-  // 1.create the cabbin
-  const { data, error } = await supabase
-    .from("cabins")
-    .insert([{ ...newCabin, image: imagePath }]);
+  // A) CREATE
+  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
+
+  // B) EDIT
+  if (id) query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
+
+  const { data, error } = await query.select().single();
 
   if (error) {
     console.error(error);
-    throw new Error("Cabin could not be inserted");
+    throw new Error("Cabin could not be created");
   }
 
-  // 2 upload image
+  // 2. Upload image
+  if (hasImagePath) return data;
+
   const { error: storageError } = await supabase.storage
-    .from("avatars")
+    .from("cabin-images")
     .upload(imageName, newCabin.image);
 
-  // 3.delete the cabin IF there was an error uplaoding image
+  // 3. Delete the cabin IF there was an error uplaoding image
   if (storageError) {
     await supabase.from("cabins").delete().eq("id", data.id);
     console.error(storageError);
@@ -50,5 +53,17 @@ export async function createCabin(newCabin) {
       "Cabin image could not be uploaded and the cabin was not created"
     );
   }
+
+  return data;
+}
+
+export async function deleteCabin(id) {
+  const { data, error } = await supabase.from("cabins").delete().eq("id", id);
+
+  if (error) {
+    console.error(error);
+    throw new Error("Cabin could not be deleted");
+  }
+
   return data;
 }
